feat: move monthlychart components to src/pages/listPage/monthlyChart
